Refactor service DTOs and update authorization
Updated `ServiceBusinessTests` to use `ServiceCreateRequestDto` instead of `ServiceUpdateRequestDto`, and changed the `Prices` property type. Modified `ServicesFunction` to allow anonymous access for service creation and update endpoints. Adjusted the `Update` method in `ServiceBusiness` to accept the new DTO. Added a `Status` property to the `ReservePrice` class and updated the `IServiceBusiness` interface accordingly. Created migration `20250507224207_InitialMigration-pt3` to add the `Status` column to the `ReservePrice` table and updated the model snapshot.